backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic
python-multipart
